fixes some small bugs, fix so that feature_memory only is added to feature models with _store_feautes == True. And added a dummy NN for debugging and so on

+++ hotam/nn/models/lstm_crf.py
@@ -92,7 +92,7 @@ class LSTM_CRF(nn.Module):
 
         lengths = batch["lengths_tok"]
         mask = batch["token_mask"]
-        word_embs = batch["word_embs"].float()
+        word_embs = batch["word_embs"]
 
         if self.use_dropout:
             word_embs = self.dropout(word_embs)
